Added beamer and login bloc and authentication bloc to auth user, switched the...

See merge request server-side-scripting-frameworks2/bookingsystem!1

+++ api/resolvers/officeResolver.js
@@ -1,25 +1,30 @@
 'use strict';
 import Office from '../models/office';
+import {checkPermission} from "../utils/auth";
 
 export default {
     Query: {
-        offices: async (parent, args) => {
-            console.log('offices', parent, args);
-            return await Office.find({organizationID: args.organizationID});
+        offices: async (parent, {organizationID}, {user}) => {
+            checkPermission(user, false);
+            console.log('offices', parent, organizationID);
+            return await Office.find({organizationID: organizationID});
         }
     },
     Mutation: {
-        addOffice: async (parent, args) => {
+        addOffice: async (parent, args, {user}) => {
+            checkPermission(user, true);
             console.log('addOffice', parent, args);
             const newOffice = new Office(args);
             console.log('newOffice', newOffice);
             return await newOffice.save();
         },
-        renameOffice: async (parent, args) => {
+        renameOffice: async (parent, args, {user}) => {
+            checkPermission(user, true);
             console.log('renameOffice', parent, args);
             return await Office.findByIdAndUpdate(args.officeID, {name: args.newName}, {new: true,});
         },
-        deleteOffice: async (parent, args) => {
+        deleteOffice: async (parent, args, {user}) => {
+            checkPermission(user, true);
             console.log('deleteOffice', parent, args);
             return await Office.findOneAndDelete({_id: args.id});
         }
